Fix for :rtype: being rendered in text
Adds a blank line before :rtype: when it is appended. This ensures that
Sphinx doesn't interpret it as part of normal text.

I've only reproduced the problem when using sphinx.ext.napoleon (just
having it loaded is enough; one doesn't need to use numpydoc
formatting), so it's added to the test suite.