Use format_annotation to render class attribute type annotations
Resolves issue #298.

This is a bit hard to do because autodoc doesn't emit any events to help us
adjust the type rendering for attributes and because the type annotation is
currently converted to text and then rendered from text by the `PyAttribute`
directive.

The system in autodoc for attribute rendering is:
1. In `AttributeDocumenter.get_directive_header`, call `stringify_annotation` to
   render the annotation to a string
2. add `:type: stringified annotation` to the directive header of `PyAttribute`
3. When `sphinx.domains.python.PyAttribute` sees the `:type:` it calls
   `format_annotation` which attempts to convert the string representation of
   the type into a more fancy representation with cross referencing

Notes:
2. we don't get an opportunity to adjust the directive header with
   `autodoc-process-docstring` or some other event
3. there isn't enough information retained in the string for
   `_inject_types_to_docstring` to work.

So my approach is:
1. Patch `AttributeDocumenter.get_directive_header` to use `format_annotation`
   instead of `stringify_annotation`. Add a prefix to the output to indicate
   that it is rst not a plain text type annotation.
3. Patch `PyAttribute` so that when it sees the prefix, it formats the output as
   rst rather than trying to handle it as a plain text type annotation.

This is a bit horrible. As future proofing, we check that the patch targets
exist and bail if they don't exist. If they move this patch should gracefully
decay into a no-op rather than crashing.