More robust determination of rtype location / fix issue 302
Resolves 302.

This increases the robustness of the handling of the docutils tree.
The docutils rst parser seems to have a bunch of bugs / shortcomings.
The original problem was that several list types do not get their line
number set. So to find the line number of a node, if the node itself has
None as the line number, we descend through the first children of the
node until we find one with a non-None line number.