Refactor tests to make it easier to add new example functions (#306)
* Refactor tests to make it easier to add more example functions

This makes it so that it is possible to add a new example while only
modifying one location in the code. Add a new class / function / exception
in test_integration, indicate the expected output with the @expected decorator.
More efficient than having to modify three different files like we had to do
before.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Some minor improvements

* Fix test

* Ignore types in test_integration.py

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>