Patch docutils rst parser to add line numbers to more nodes (#316)
* Patch docutils rst parser to add line numbers to more nodes

This resolves issue 312. The missing line numbers make it impossible
for us to correctly calculate where to place the rtype.

See upstream issue:
https://sourceforge.net/p/docutils/bugs/465/

I submitted a patch to docutils to fix these things but it is currently
waiting to get through a spam filter. The docutils test suite has no
coverage checking whether this field is filled correctly.

* Remove noqa for spelling error